# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ResilAI",
    "description": "ResilAI - AI Incident Readiness Score API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/clinic/readiness/{org_id}": {
      "get": {
        "tags": [
          "clinic"
        ],
        "summary": "Get Clinic Readiness",
        "description": "The product endpoint. Returns the immutable DailyReadinessReport.",
        "operationId": "get_clinic_readiness_api_clinic_readiness__org_id__get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReadinessReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/clinic/problems/{problem_id}/fix": {
      "post": {
        "tags": [
          "clinic"
        ],
        "summary": "Fix Problem",
        "description": "Triggers the autofix for a given problem.",
        "operationId": "fix_problem_api_clinic_problems__problem_id__fix_post",
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Problem Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionCard": {
        "properties": {
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "why_it_matters": {
            "type": "string",
            "title": "Why It Matters"
          },
          "recommended_action": {
            "type": "string",
            "title": "Recommended Action"
          },
          "can_be_undone": {
            "type": "boolean",
            "title": "Can Be Undone"
          },
          "estimated_time_minutes": {
            "type": "integer",
            "title": "Estimated Time Minutes"
          },
          "fix_now_available": {
            "type": "boolean",
            "title": "Fix Now Available"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "approval_needed": {
            "type": "boolean",
            "title": "Approval Needed"
          },
          "required_permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Permissions"
          },
          "success_message": {
            "type": "string",
            "title": "Success Message"
          }
        },
        "type": "object",
        "required": [
          "action_id",
          "problem",
          "why_it_matters",
          "recommended_action",
          "can_be_undone",
          "estimated_time_minutes",
          "fix_now_available",
          "category",
          "approval_needed",
          "required_permissions",
          "success_message"
        ],
        "title": "ActionCard",
        "description": "Customer-facing action. Replaces raw ActionIntent at the product boundary."
      },
...
      "BusinessContinuity": {
        "properties": {
          "operational_readiness": {
            "$ref": "#/components/schemas/OperationalReadiness"
          }
        },
        "type": "object",
        "required": [
          "operational_readiness"
        ],
        "title": "BusinessContinuity",
        "description": "Top-level container for continuity and operational metrics."
      },
...
      "ConnectorReadiness": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "connected": {
            "type": "boolean",
            "title": "Connected"
          },
          "last_verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At"
          },
          "coverage": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Coverage"
          },
          "missing_visibility": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Visibility"
          },
          "confidence_pct": {
            "type": "integer",
            "title": "Confidence Pct",
            "default": 100
          }
        },
        "type": "object",
        "required": [
          "name",
          "status",
          "connected"
        ],
        "title": "ConnectorReadiness",
        "description": "Health and coverage of one integration."
      },
      "CoverageArea": {
        "properties": {
          "area": {
            "type": "string",
            "title": "Area"
          },
          "monitored": {
            "type": "boolean",
            "title": "Monitored"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "area",
          "monitored"
        ],
        "title": "CoverageArea",
        "description": "One security area that can be monitored."
      },
      "CoverageReport": {
        "properties": {
          "coverage_pct": {
            "type": "integer",
            "title": "Coverage Pct"
          },
          "monitored": {
            "items": {
              "$ref": "#/components/schemas/CoverageArea"
            },
            "type": "array",
            "title": "Monitored"
          },
          "not_monitored": {
            "items": {
              "$ref": "#/components/schemas/CoverageArea"
            },
            "type": "array",
            "title": "Not Monitored"
          }
        },
        "type": "object",
        "required": [
          "coverage_pct"
        ],
        "title": "CoverageReport",
        "description": "Security Visibility: what we can see and what we can't."
      },
      "DailyReadinessReport": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "report_date": {
            "type": "string",
            "title": "Report Date"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "status": {
            "$ref": "#/components/schemas/ReadinessStatus"
          },
          "clinic_health_pct": {
            "type": "integer",
            "title": "Clinic Health Pct"
          },
          "connector_health_pct": {
            "type": "integer",
            "title": "Connector Health Pct"
          },
          "greeting": {
            "type": "string",
            "title": "Greeting"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "type": "array",
            "title": "Timeline"
          },
          "business_continuity": {
            "$ref": "#/components/schemas/BusinessContinuity"
          },
          "passed_checks": {
            "items": {
              "$ref": "#/components/schemas/ReadinessCheck"
            },
            "type": "array",
            "title": "Passed Checks"
          },
          "failed_checks": {
            "items": {
              "$ref": "#/components/schemas/ReadinessCheck"
            },
            "type": "array",
            "title": "Failed Checks"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ReadinessCheck"
            },
            "type": "array",
            "title": "Warnings"
          },
          "unknowns": {
            "items": {
              "$ref": "#/components/schemas/UnknownItem"
            },
            "type": "array",
            "title": "Unknowns"
          },
          "immediate_actions": {
            "items": {
              "$ref": "#/components/schemas/ActionCard"
            },
            "type": "array",
            "title": "Immediate Actions"
          },
          "coverage": {
            "$ref": "#/components/schemas/CoverageReport"
          },
          "connectors": {
            "items": {
              "$ref": "#/components/schemas/ConnectorReadiness"
            },
            "type": "array",
            "title": "Connectors"
          },
          "verification": {
            "$ref": "#/components/schemas/VerificationExplanation"
          },
          "trend": {
            "$ref": "#/components/schemas/ReadinessTrend"
          },
          "value": {
            "$ref": "#/components/schemas/ValueSummary"
          },
          "audit_snapshot_id": {
            "type": "string",
            "title": "Audit Snapshot Id",
            "default": ""
          },
          "checks_performed": {
            "type": "integer",
            "title": "Checks Performed",
            "default": 0
          },
          "devices_checked": {
            "type": "integer",
            "title": "Devices Checked",
            "default": 0
          },
          "accounts_checked": {
            "type": "integer",
            "title": "Accounts Checked",
            "default": 0
          },
          "backups_verified": {
            "type": "integer",
            "title": "Backups Verified",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "report_date",
          "generated_at",
          "status",
          "clinic_health_pct",
          "connector_health_pct",
          "greeting",
          "summary",
          "business_continuity"
        ],
        "title": "DailyReadinessReport",
        "description": "The immutable product contract.\n\nEverything in the system \u2014 every connector, capability, evidence provider,\nbusiness rule, action, and trust signal \u2014 exists ONLY to populate this object.\n\nThis is what the customer sees. This is what they pay for.\nThis is what answers: \"Can this clinic safely open today?\""
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperationalReadiness": {
        "properties": {
          "can_operate_today": {
            "type": "boolean",
            "title": "Can Operate Today"
          },
          "can_recover": {
            "type": "boolean",
            "title": "Can Recover"
          },
          "current_blockers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Current Blockers"
          },
          "estimated_downtime_minutes": {
            "type": "integer",
            "title": "Estimated Downtime Minutes",
            "default": 0
          },
          "critical_systems_verified": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Critical Systems Verified"
          },
          "critical_systems_assumed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Critical Systems Assumed"
          }
        },
        "type": "object",
        "required": [
          "can_operate_today",
          "can_recover"
        ],
        "title": "OperationalReadiness",
        "description": "Answers core business continuity questions."
      },
...
      "ReadinessCheck": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionCard"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status",
          "label"
        ],
        "title": "ReadinessCheck",
        "description": "One item in the readiness report. Includes verification and optional action."
      },
      "ReadinessStatus": {
        "type": "string",
        "enum": [
          "safe_to_open",
          "action_needed",
          "critical_risk",
          "unknown"
        ],
        "title": "ReadinessStatus",
        "description": "The three states a clinic can be in."
      },
      "ReadinessTrend": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/ReadinessTrendPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "improving": {
            "type": "boolean",
            "title": "Improving",
            "default": true
          }
        },
        "type": "object",
        "title": "ReadinessTrend",
        "description": "Are we improving? Historical daily snapshots with explanations."
      },
      "ReadinessTrendPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "clinic_health_pct": {
            "type": "integer",
            "title": "Clinic Health Pct"
          },
          "delta_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Delta Reasons"
          }
        },
        "type": "object",
        "required": [
          "date",
          "clinic_health_pct"
        ],
        "title": "ReadinessTrendPoint",
        "description": "One day's readiness snapshot."
      },
...
      "TimelineEvent": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "icon",
          "label",
          "source"
        ],
        "title": "TimelineEvent",
        "description": "One entry in the overnight timeline. Chronological."
      },
      "UnknownItem": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "last_seen_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen Description"
          },
          "impact": {
            "type": "string",
            "title": "Impact"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "label",
          "impact",
          "source"
        ],
        "title": "UnknownItem",
        "description": "Something we can't verify. 'We don't know' is more trustworthy than silence."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValueSummary": {
        "properties": {
          "period_label": {
            "type": "string",
            "title": "Period Label"
          },
          "accounts_protected": {
            "type": "integer",
            "title": "Accounts Protected",
            "default": 0
          },
          "devices_protected": {
            "type": "integer",
            "title": "Devices Protected",
            "default": 0
          },
          "backups_verified": {
            "type": "integer",
            "title": "Backups Verified",
            "default": 0
          },
          "problems_prevented": {
            "type": "integer",
            "title": "Problems Prevented",
            "default": 0
          },
          "estimated_downtime_avoided_hours": {
            "type": "number",
            "title": "Estimated Downtime Avoided Hours",
            "default": 0.0
          },
          "estimated_hipaa_records_protected": {
            "type": "integer",
            "title": "Estimated Hipaa Records Protected",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "period_label"
        ],
        "title": "ValueSummary",
        "description": "Business value delivered. These are what convince customers to renew."
      },
...
      "VerificationContext": {
        "properties": {
          "verification_source": {
            "type": "string",
            "title": "Verification Source"
          },
          "last_verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At"
          },
          "connector_health": {
            "type": "string",
            "title": "Connector Health"
          },
          "confidence_pct": {
            "type": "integer",
            "title": "Confidence Pct"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "data_age_description": {
            "type": "string",
            "title": "Data Age Description"
          },
          "can_reverify": {
            "type": "boolean",
            "title": "Can Reverify",
            "default": true
          },
          "verification_method": {
            "type": "string",
            "title": "Verification Method"
          }
        },
        "type": "object",
        "required": [
          "verification_source",
          "connector_health",
          "confidence_pct",
          "verification_status",
          "data_age_description",
          "verification_method"
        ],
        "title": "VerificationContext",
        "description": "Per-item verification metadata. Answers 'Why are you telling me this?'"
      },
      "VerificationExplanation": {
        "properties": {
          "confidence_pct": {
            "type": "integer",
            "title": "Confidence Pct"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/VerificationReason"
            },
            "type": "array",
            "title": "Reasons"
          }
        },
        "type": "object",
        "required": [
          "confidence_pct"
        ],
        "title": "VerificationExplanation",
        "description": "Why we believe our assessment. Doctors trust explanations, not numbers."
      },
      "VerificationReason": {
        "properties": {
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "icon",
          "text"
        ],
        "title": "VerificationReason",
        "description": "One reason contributing to verification or distrust."
      }
    }
  }
}